refactor(phase-3D): split agent/loop.py into message_utils + tool_manager + guidance [plan-deletion]
Phase 3D / agent/REFACTORING.md §3 Step 2.

``ari/agent/loop.py`` shrinks from 1459 → 1246 lines.  The ``run()``
body itself stays put per the plan ("run() 自体を更に分割するかは
Phase 3D の本 PR ではやらない"); three helper clusters move to
sibling modules with the class methods staying as 1-line delegators
so subclasses + monkeypatches keep working byte-for-byte.

New modules:
- ``ari/agent/message_utils.py`` — ``_extract_job_ids`` +
  ``_tool_was_called``.  loop.py keeps the original names via a
  re-export ``from ari.agent.message_utils import ...``.
- ``ari/agent/tool_manager.py`` — ``available_tools_openai`` /
  ``execute_tool_calls`` / ``active_tools``; the class methods
  ``_available_tools_openai`` / ``_execute_tool_calls`` /
  ``_active_tools`` are 1-line delegators.
- ``ari/agent/guidance.py`` — ``guidance`` + ``validate_metrics``;
  ``_guidance`` / ``_validate_metrics`` methods delegate.

Verified before plan deletion (agent §4 + §0 contract):

- ✅ ``test_agent_smoke.py`` (12 cases) green — Phase-0 regression
  net catches helper drift.
- ✅ ``test_react_driver.py`` (18 cases) green.
- ✅ ``test_run_loop.py`` (12 cases) green.
- ✅ ``test_disabled_tools_flow.py`` (6 cases) green.
- ✅ ``test_prompt_extraction.py`` agent/system sha256 unchanged
  (PC3 byte-identical lift).
- ✅ ``AgentLoop.__init__`` / ``run`` public signatures unchanged.
- ✅ ``WorkflowHints`` dataclass unchanged.
- ✅ Phase PC3 SYSTEM_PROMPT extraction byte-identical (sha256
  ``a50abe13d568c0...``).

Permanent knowledge transcribed:
- §1 file responsibility table → ``docs/refactor_audit.md`` already
  carries the pre-split sizes; the post-split layout is
  self-documenting via each new module's docstring.
- §0 BFTS determinism contract → ``docs/PHILOSOPHY.md`` covers P2.

pytest ari-core/tests/ -q (excluding live integrations):
1971 passed, 15 skipped.
